MOVE: move to pandas/core/describe.py
