+++ pandas/core/describe.py
